FIX: SVG doesn't need to be direct child of toggle
Removed unnecessary ">" from svg selector.
Thus, the actual toggle (element holding the class representing folded or unfolded state) does not have to be the parent of the inserted SVG (i.e. the element on which the pre|appendFoldingArrowIcon() plug-in has been called), but the transform/setupTransformation methods may also be called on an earlier ancestor).
Also minor changes to README.