Added width and height attributes
The inline SVGs are now fitted with width and height attributes. As long as the CSS styles are loaded, this does have no effect at all. But in case the page loads without stylesheet(s), now the (totally unstyled) icons are relatively small and not scaled up to fit the window width like before. This should improve accessibility for users who use a site without CSS or with custom CSS.